requirements.txt
# Core dependencies for PDF processing and vector database
pdfplumber>=0.9.0
pytesseract>=0.3.10
Pillow>=10.0.0

# LangChain for document processing and embeddings
langchain>=0.1.0
langchain-community>=0.0.13
langchain-huggingface>=0.0.3

# Vector database
chromadb>=0.4.0

# Embeddings
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
ollama>=0.1.0
huggingface-hub>=0.16.0
tiktoken>=0.4.0

# NumPy - Fixed version for compatibility
numpy>=1.24.0,<2.0.0

# Additional dependencies
typing-extensions>=4.0.0